requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
datasets>=2.15.0
accelerate>=0.25.0
peft>=0.6.0               # Parameter-Efficient Fine-Tuning (LoRA, etc.)
bitsandbytes>=0.41.0      # 4-bit / 8-bit quantization

# Tokenization
sentencepiece>=0.1.99
tokenizers>=0.15.0

# Evaluation metrics
evaluate>=0.4.0
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.1
nltk>=3.8.1

# Training utilities
tensorboard>=2.14.0
wandb>=0.16.0
tqdm>=4.66.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization (notebooks)
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
ipywidgets>=8.1.0

# Notebook support
jupyter>=1.0.0
ipykernel>=6.27.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
